Fix OOM and simplify feature extraction for Rorqual cluster
- Use 2g.20gb GPU slice (20GB VRAM) instead of 1g.10gb to fix LLAMA OOM
- Rewrite run_feature_extraction.sh as interactive salloc script with configurable time
- Reduce test subset to subject_timeline_index<1 for faster iteration
- Add h5py dep, BFloat16 fix, HF offline mode, cluster setup docs, diagnose script

+++ data_utils/data_utils/features/text.py
@@ -252,5 +252,5 @@ class LLAMA3p2(pydantic.BaseModel):
                 word_state = hidden_state[:, -len(target_word) :]
 
                 word_state = word_state.mean(axis=1)
-                out = word_state.cpu().numpy()
+                out = word_state.cpu().float().numpy()
                 yield out
